Custom expression: switch to a recursive-descent parser
The fault-tolerant/recovery mode of the Chevrotain-based parser isn't
needed anymore since the suggestion was made to be lexical-based in
PR #15688. Therefore, there is no more reason to carry the Chevrotain
baggage. The grammar construct for Chevrotain was converted into a
simple recursive-descen parser. This represents a reduction from over
1,100 LoC to a leaner parser of only 270 LoC.

Also, we remove the dependency to Chevrotain, reducing the front-end
bundle by approx 18 KB.